requirement.txt
flask==2.3.2
pinecone-client==2.2.1